smoothed out animation for options drop down.
