+++ src/index.scss
@@ -194,14 +194,18 @@ a {
 }
 
 .bounder {
+  border-radius: 5px;
+  transition: height 0.5s;
   &-enter-active {
     overflow: hidden;
+    box-shadow: 0 1px 4px 0 rgb(0 0 0 / 37%);
   }
   &-enter-done {
     overflow: visible;
   }
   &-exit-active {
     overflow: hidden;
+    box-shadow: 0 1px 4px 0 rgb(0 0 0 / 37%);
   }
 }
